Complete Add Two Numbers Problem
It is classified as Medium under link linked list section. I was able to
solve it on time by myself. My personal solution uses strings to get
the values from the total. The better solution is to use floor division
and modulo to get the values from total. It is not listed in
need-more-prac.txt.