Complete Linked List Cycle Problem
It is classified as Easy and is under linked list section. I was able
to solve it within 5 minutes with an suboptimal memory solution. My
solution utilizes a hashmaps to check if a node is visited. A better
solution uses two pointers, specifically, slow and fast. A loop exists
if fast pointer reached the slow pointer. If fast pointer reaches null
then it is not a cycle linked list. Both solution are included. This
problem is listed to need-more-prac.txt to practice the optimal
solution.