Practice Reorder List
It is classified as easy and is under linked list. I was able to solve
it using an array for each node which takes O(n) memory and is not
optimized. The better solution would be split the list in half, reverse
the second half, then merge both halves. '(1)' is added in
need-more-prac.txt to indicate that it took me two tries already. It is
not removed from the txt file.