Practice Remove Nth Node from End of the List
It is classified as medium and is under linked list. I was able to solve
it using an array for each node which takes O(n) memory and is not
optimized. The better solution would be to use two pointer with gap n.
If end pointer reach None, then start pointer is in the previous node of
the node to be removed. It is not added to need-more-prac.txt.