Practice Merge k Sorted Lists Problem
It is classified as hard and is under linked list. I was able to come up
with a solution and solve it within the 45 mins time limit. However my
solution uses O(n * m) time which is not optimized. A better solution
would be to use a merge sort algorithm (Sort two list per iteration,
then merge them together until one list remain). In need-more-prac.txt,
the problem is indicated with '(1)' to show many failed tries were done.
It is not removed from need-more-prac.txt.